microservice/detect-and-teach-nail-art/requirements.txt
fastapi
uvicorn
openai
requests
Pillow
python-dotenv
python-multipart